requirements.txt
python-telegram-bot
selenium
webdriver_manager